+++ config/hard/Tetris/code/CGCode.java
@@ -1,0 +1,31 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        int SW = in.nextInt();
+        int SH = in.nextInt();
+        if (in.hasNextLine()) {
+            in.nextLine();
+        }
+        for (int i = 0; i < SH; i++) {
+            String SROW = in.nextLine();
+        }
+        int FW = in.nextInt();
+        int FH = in.nextInt();
+        if (in.hasNextLine()) {
+            in.nextLine();
+        }
+        for (int i = 0; i < FH; i++) {
+            String FROW = in.nextLine();
+        }
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("answer");
+    }
+}