+++ config/hard/Tetris/code/CGCode.js
@@ -1,0 +1,17 @@
+var inputs = readline().split(' ');
+const SW = parseInt(inputs[0]);
+const SH = parseInt(inputs[1]);
+for (let i = 0; i < SH; i++) {
+    const SROW = readline();
+}
+var inputs = readline().split(' ');
+const FW = parseInt(inputs[0]);
+const FH = parseInt(inputs[1]);
+for (let i = 0; i < FH; i++) {
+    const FROW = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');