+++ config/hard/Tetris/code/CGCode.ts
@@ -1,0 +1,17 @@
+var inputs: string[] = readline().split(' ');
+const SW: number = parseInt(inputs[0]);
+const SH: number = parseInt(inputs[1]);
+for (let i = 0; i < SH; i++) {
+    const SROW: string = readline();
+}
+var inputs: string[] = readline().split(' ');
+const FW: number = parseInt(inputs[0]);
+const FH: number = parseInt(inputs[1]);
+for (let i = 0; i < FH; i++) {
+    const FROW: string = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');