+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Shape Recognition".
 - configuration for "n Queens".
 - configuration for "Horn-SAT solver".
+- configuration for "Tetris".
 
 ## [1.15.0] - 2025-05-31
 ### Added